% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citippy.R
\name{citippy}
\alias{citippy}
\title{add citation to rmarkdown document that will show as tooltip on hover}
\usage{
citippy(link, bibdf = bibdf, bibref = bibref, pandoc = params$pandoc,
  cite_style = "numeric", textual = FALSE, ...)
}
\arguments{
\item{link}{Bibtex key}

\item{bibdf}{dataframe of bibentries returned by bib2df::bib2df}

\item{bibref}{bibentries returned by RefManageR::ReadBib}

\item{pandoc}{Use pandoc to handle the citations and references (TRUE) or 
display citations with tooltips showing references on hover (FALSE) in html 
documents. Default is FALSE.}

\item{cite_style}{Options for styling of in-text citations. Options are 
"numeric", "authoryear". Default is set to "numeric".}

\item{...}{Additional arguments to pass to `tippy::tippy`}
}
\description{
add citation to rmarkdown document that will show as tooltip on hover
}
