% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citippy_path.R
\name{citippy_path}
\alias{citippy_path}
\title{add citation to rmarkdown document that will show as tooltip on hover}
\usage{
citippy_path(ref_path)
}
\arguments{
\item{ref_path}{Path to your .bib file with bibtex entries}
}
\description{
add citation to rmarkdown document that will show as tooltip on hover
}
