% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citippy_ref.R
\name{citippy_ref}
\alias{citippy_ref}
\title{use crossref to find a citation and add to references.bib file}
\usage{
citippy_ref(query, ref_path = "references.bib")
}
\arguments{
\item{query}{Search string}

\item{ref_path}{Default is references.bib}
}
\description{
use crossref to find a citation and add to references.bib file
}
