% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citippy.R
\name{citippy}
\alias{citippy}
\title{citippy}
\usage{
citippy(link, ref = "references.bib", pandoc = params$pandoc,
  cite_style = "numeric")
}
\arguments{
\item{ref}{path to .bib reference file}

\item{pandoc}{logical but best to set as parameter in
rmarkdown yaml}

\item{cite_style}{character string; bibliography style to use to
generate citations.}
}
\description{
citippy
}
